refactor: expose formatting helpers
